Split HotwireLogger into focused files under Logging/
Separate the public logging API (LogLevel, Logger protocol + defaults,
OSLogWrapper) from the internal wiring (HotwireLogger + global logger),
one type per file. Move the LogLevel->OSLogType mapping into
OSLogWrapper so LogLevel no longer depends on OSLog.
