+++ Source/Logging/LogLevel.swift
@@ -1,0 +1,6 @@
+public enum LogLevel {
+    case debug
+    case info
+    case warning
+    case error
+}
